fix dnd and cards refactor
